wip(move): Return MemberCaps instead of transferring them ourselves. Get rid of the EncryptionKey, instead use a vector<vector<u8>> on the Channel, to maintain a history of the channel keys. TODO1: modify the ts-sdk to account for the contract changes. TODO2: finish the CreateChannelFlow TODO3: write integration tests

+++ move/sui_messaging/sources/attachment.move
@@ -13,7 +13,7 @@ public struct Attachment has copy, drop, store {
     encrypted_metadata: vector<u8>,
     data_nonce: vector<u8>, // Need separate nonces, because we must not reuse a nonce when encrypting with the same key
     metadata_nonce: vector<u8>,
-    key_version: u32,
+    key_version: u64,
 }
 
 // === Events ===
@@ -26,7 +26,7 @@ public fun new(
     encrypted_metadata: vector<u8>,
     data_nonce: vector<u8>,
     metadata_nonce: vector<u8>,
-    key_version: u32,
+    key_version: u64,
 ): Attachment {
     Attachment {
         blob_ref,
